+++ src/main/java/org/pushpmx/problem/Nguyen5Problem.java
@@ -3,8 +3,8 @@ package org.pushpmx.problem;
 public class Nguyen5Problem extends FloatSymbolicRegression {
 
 	@Override
-	protected float evaluateFunction(float x) {
-		return (float) (Math.sin(x * x) * Math.cos(x) - 1.0f);
+	protected double evaluateFunction(double x) {
+		return (Math.sin(x * x) * Math.cos(x) - 1.0);
 	}
 
 }
