+++ src/main/java/org/pushpmx/problem/R1Problem.java
@@ -3,7 +3,7 @@ package org.pushpmx.problem;
 public class R1Problem extends FloatSymbolicRegression {
 
 	@Override
-	protected float evaluateFunction(float x) {
+	protected double evaluateFunction(double x) {
 		return (x + 1) * (x + 1) * (x + 1) / (x * x - x + 1);
 	}
